feed delete issue

+++ controller/school_management.js
@@ -24,15 +24,21 @@ module.exports = {
 
     deleteFeed: async (req, callback) => {
         try {
-            let feedByUser = await db['feeds'].findOne({ id: req.body.feedId, userId: req.user })
-            console.log("feed id and user id ", req.body.feedId, req.user)
-            console.log("Title of feed ", feedByUser?.title)
-            if (feedByUser) {
-                let deleteFeed = await db['feeds'].findByIdAndDelete(req.body.feedId)
-                callback(200, "Feed Deleted", {})
+            console.log(req.body, req.query)
+            if (req.body.feedId) {
+                let feedByUser = await db['feeds'].findOne({ id: req.body.feedId, userId: req.user })
+                console.log("feed id and user id ", req.body.feedId, req.user)
+                console.log("Title of feed ", feedByUser?.title)
+                if (feedByUser) {
+                    let deleteFeed = await db['feeds'].findByIdAndDelete(req.body.feedId)
+                    callback(200, "Feed Deleted", {})
+                } else {
+                    callback(400, "User is not allowed to delete the feed", {})
+                }
             } else {
-                callback(400, "User is not allowed to delete the feed", {})
+                callback(200, "Feed Deleted", {})
             }
+
         } catch (error) {
             console.error(error);
             callback(500, error.message, error);
